requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.3.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=1.11.0=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.13.1=pypi_0
aiosignal=1.4.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.11.0=pypi_0
async-timeout=5.0.1=pypi_0
attrs=25.4.0=pypi_0
blas=1.0=mkl
brotlicffi=1.0.9.2=py310h6a678d5_1
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.9.9=h06a4308_0
certifi=2025.10.5=py310h06a4308_0
cffi=2.0.0=py310h4eded50_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.3.0=pypi_0
cuda-cudart=12.4.127=0
cuda-cupti=12.4.127=0
cuda-libraries=12.4.1=0
cuda-nvrtc=12.4.127=0
cuda-nvtx=12.4.127=0
cuda-opencl=12.9.19=0
cuda-runtime=12.4.1=0
cuda-version=12.9=3
datasets=4.2.0=pypi_0
diffusers=0.21.0=pypi_0
dill=0.4.0=pypi_0
einops=0.8.1=pypi_0
exceptiongroup=1.3.0=pypi_0
expat=2.7.1=h6a678d5_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.17.0=py310h06a4308_0
freetype=2.13.3=h4a9f257_0
frozenlist=1.8.0=pypi_0
fsspec=2025.9.0=pypi_0
ftfy=6.3.1=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.45=pypi_0
gmp=6.3.0=h6a678d5_0
gmpy2=2.2.1=py310h5eee18b_0
gnutls=3.6.15=he1e5248_0
h11=0.16.0=pypi_0
hf-xet=1.1.10=pypi_0
httpcore=1.0.9=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.35.3=pypi_0
idna=3.7=py310h06a4308_0
imageio=2.37.0=pypi_0
importlib-metadata=8.7.0=pypi_0
intel-openmp=2023.0.0=h9e868ea_25371
jinja2=3.1.6=py310h06a4308_0
joblib=1.5.2=pypi_0
jpeg=9f=h5ce9db8_0
lame=3.100=h7b6447c_0
lazy-loader=0.4=pypi_0
lcms2=2.16=h92b89f2_1
ld_impl_linux-64=2.44=h153f514_2
lerc=4.0.0=h6a678d5_0
libcublas=12.4.5.8=0
libcufft=11.2.1.3=0
libcufile=1.14.1.1=4
libcurand=10.3.10.19=0
libcusolver=11.6.1.9=0
libcusparse=12.3.1.170=0
libdeflate=1.22=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=12.2.5.30=0
libnvfatbin=12.9.82=0
libnvjitlink=12.4.127=0
libnvjpeg=12.3.1.117=0
libopus=1.3.1=h5eee18b_1
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.20.0=ha30cf84_0
libtiff=4.7.0=hde9077f_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libvpx=1.7.0=h439df22_0
libwebp-base=1.3.2=h5eee18b_1
libxcb=1.17.0=h9b100fa_0
libzlib=1.3.1=hb25bd0a_0
llvm-openmp=14.0.6=h9e868ea_0
lpips=0.1.4=pypi_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=3.0.2=py310h5eee18b_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_2
mkl_fft=1.3.11=py310h5eee18b_0
mkl_random=1.2.8=py310h1128e8f_0
mpc=1.3.1=h5eee18b_0
mpfr=4.2.1=h5eee18b_0
mpmath=1.3.0=py310h06a4308_0
multidict=6.7.0=pypi_0
multiprocess=0.70.16=pypi_0
ncurses=6.5=h7934f7d_0
nettle=3.7.3=hbbd107a_1
networkx=3.4.2=py310h06a4308_0
numpy=2.0.1=py310h5f9d8c6_1
numpy-base=2.0.1=py310hb5e798b_1
nvidia-cublas-cu12=12.8.4.1=pypi_0
nvidia-cuda-cupti-cu12=12.8.90=pypi_0
nvidia-cuda-nvrtc-cu12=12.8.93=pypi_0
nvidia-cuda-runtime-cu12=12.8.90=pypi_0
nvidia-cudnn-cu12=9.10.2.21=pypi_0
nvidia-cufft-cu12=11.3.3.83=pypi_0
nvidia-cufile-cu12=1.13.1.3=pypi_0
nvidia-curand-cu12=10.3.9.90=pypi_0
nvidia-cusolver-cu12=11.7.3.90=pypi_0
nvidia-cusparse-cu12=12.5.8.93=pypi_0
nvidia-cusparselt-cu12=0.7.1=pypi_0
nvidia-nccl-cu12=2.27.5=pypi_0
nvidia-nvjitlink-cu12=12.8.93=pypi_0
nvidia-nvshmem-cu12=3.3.20=pypi_0
nvidia-nvtx-cu12=12.8.90=pypi_0
ocl-icd=2.3.2=h5eee18b_1
open-clip-torch=3.2.0=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=h0d4d230_1
openssl=3.0.18=hd6dcaed_0
packaging=25.0=pypi_0
pandas=2.3.3=pypi_0
pillow=11.1.0=py310hac6e08b_1
pip=25.2=pyhc872135_1
platformdirs=4.5.0=pypi_0
propcache=0.4.1=pypi_0
protobuf=6.33.0=pypi_0
psutil=7.1.1=pypi_0
pthread-stubs=0.3=h0ce48e5_1
pyarrow=21.0.0=pypi_0
pycparser=2.23=py310h06a4308_0
pydantic=2.12.3=pypi_0
pydantic-core=2.41.4=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.18=h1a3bd86_0
python-dateutil=2.9.0.post0=pypi_0
pytorch=2.4.1=py3.10_cuda12.4_cudnn9.1.0_0
pytorch-cuda=12.4=hc786d27_7
pytorch-mutex=1.0=cuda
pytorch-wavelets=1.3.0=pypi_0
pytz=2025.2=pypi_0
pywavelets=1.8.0=pypi_0
pyyaml=6.0.2=py310h5eee18b_0
readline=8.3=hc2a1206_0
regex=2025.9.18=pypi_0
requests=2.32.5=py310h06a4308_0
safetensors=0.6.2=pypi_0
scikit-image=0.25.2=pypi_0
scikit-learn=1.7.2=pypi_0
scipy=1.15.3=pypi_0
sentry-sdk=2.42.1=pypi_0
setuptools=80.9.0=py310h06a4308_0
six=1.17.0=pypi_0
smmap=5.0.2=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.50.2=hb25bd0a_1
sympy=1.14.0=py310h06a4308_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.6.0=pypi_0
tifffile=2025.5.10=pypi_0
timm=1.0.20=pypi_0
tk=8.6.15=h54e0aa7_0
tokenizers=0.21.4=pypi_0
torch-fidelity=0.3.0=pypi_0
torchaudio=2.4.1=py310_cu124
torchtriton=3.0.0=py310
torchvision=0.19.1=py310_cu124
tqdm=4.67.1=pypi_0
transformers=4.55.4=pypi_0
triton=3.4.0=pypi_0
typing-inspection=0.4.2=pypi_0
typing_extensions=4.15.0=py310h06a4308_0
tzdata=2025.2=pypi_0
urllib3=2.5.0=py310h06a4308_0
wandb=0.22.2=pypi_0
wcwidth=0.2.14=pypi_0
wheel=0.45.1=py310h06a4308_0
x264=1!157.20191217=h7b6447c_0
xformers=0.0.28.post1=pypi_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xxhash=3.6.0=pypi_0
xz=5.6.4=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.22.0=pypi_0
zipp=3.23.0=pypi_0
zlib=1.3.1=hb25bd0a_0
zstd=1.5.7=h11fc155_0
